Make integral promotion explicit

+++ src/containers/unrolledlist.d
@@ -508,7 +508,7 @@ private:
 		size_t nextAvailableIndex() const nothrow pure @safe @nogc
 		{
 			import containers.internal.backwards : bsf;
-			return bsf(~registry);
+			return bsf(~size_t(registry));
 		}
 
 		void markUsed(size_t index) nothrow pure @safe @nogc
